requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mysql-connector==2.2.9
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
ping3==4.0.8
platformdirs==4.2.1
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yarg==0.1.9
